feat: try to add authentik management with terraform
This changeset is fairly large in that it introduces several artifacts:

* a terraform workspace for managing authentik
* a terraform resource for management by tf-controller

This commit introduces Terraform configuration for managing the
authentik service. It includes the definition of the authentik Terraform
resource, as well as the necessary configurations for the Grafana
application within authentik.

The authentik service configuration includes OIDC client management,
bitwarden integration for storing credentials, and Grafana application
setup with specific groups.

Refs: #22